# Conflicts:
#	src/main/java/xyz/tomorrowlearncamp/bookking/domain/user/auth/controller/AuthController.java
#	src/main/java/xyz/tomorrowlearncamp/bookking/domain/user/controller/UserController.java
#	src/main/java/xyz/tomorrowlearncamp/bookking/domain/user/service/UserService.java

+++ src/main/java/xyz/tomorrowlearncamp/bookking/domain/book/controller/BookController.java
@@ -1,23 +1,14 @@
 package xyz.tomorrowlearncamp.bookking.domain.book.controller;
 
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.util.ObjectUtils;
+import org.springframework.web.bind.annotation.*;
 import xyz.tomorrowlearncamp.bookking.domain.book.dto.request.AddBookRequestDto;
 import xyz.tomorrowlearncamp.bookking.domain.book.dto.request.SearchBookRequestDto;
 import xyz.tomorrowlearncamp.bookking.domain.book.dto.request.UpdateBookRequestDto;
@@ -26,6 +17,11 @@ import xyz.tomorrowlearncamp.bookking.domain.book.dto.response.BookResponseDto;
 import xyz.tomorrowlearncamp.bookking.domain.book.dto.response.SearchBookResponseDto;
 import xyz.tomorrowlearncamp.bookking.domain.book.service.BookService;
 import xyz.tomorrowlearncamp.bookking.domain.book.service.SearchBookService;
+import xyz.tomorrowlearncamp.bookking.domain.common.dto.Response;
+import xyz.tomorrowlearncamp.bookking.domain.common.enums.ErrorMessage;
+import xyz.tomorrowlearncamp.bookking.domain.common.exception.ForbiddenRequestException;
+import xyz.tomorrowlearncamp.bookking.domain.user.auth.dto.AuthUser;
+import xyz.tomorrowlearncamp.bookking.domain.user.enums.UserRole;
 
 @RestController
 @RequiredArgsConstructor
@@ -35,60 +31,83 @@ public class BookController {
 	private final SearchBookService searchBookService;
 
 	@PostMapping("/v1/books/search")
-	@ResponseStatus(HttpStatus.OK)
-	public SearchBookResponseDto searchBooks(@RequestBody SearchBookRequestDto requestDto) {
-		return searchBookService.searchBooks(requestDto);
+	public Response<SearchBookResponseDto> searchBooks(@RequestBody SearchBookRequestDto requestDto) {
+		return Response.success(searchBookService.searchBooks(requestDto));
 	}
 
 	@PostMapping("/v1/books/import")
-	public ResponseEntity<Void> importBooks(
-		@RequestParam(defaultValue = "100") int pageSize,
-		@RequestParam(defaultValue = "100") int totalPage
+	@ResponseStatus(HttpStatus.OK)
+	public void importBooks(
+			@AuthenticationPrincipal AuthUser user,
+			@RequestParam(defaultValue = "100") int pageSize,
+			@RequestParam(defaultValue = "100") int totalPage
 	){
+		if (!ObjectUtils.nullSafeEquals(user.getRole(), UserRole.ROLE_ADMIN)) {
+			throw new ForbiddenRequestException(ErrorMessage.FORBIDDEN_ADMINISTRATOR);
+		}
 		searchBookService.fetchBooksFromLibraryApi(pageSize, totalPage);
-		return ResponseEntity.ok().build();
 	}
 
 	@PostMapping("/v1/books")
-	public ResponseEntity<Long> addBook(@RequestBody @Valid AddBookRequestDto requestDto) {
-		return ResponseEntity.ok(bookService.addBook(requestDto));
+	public Response<Long> addBook(
+			@AuthenticationPrincipal AuthUser user,
+			@RequestBody @Valid AddBookRequestDto requestDto
+	) {
+		if (!ObjectUtils.nullSafeEquals(user.getRole(), UserRole.ROLE_ADMIN)) {
+			throw new ForbiddenRequestException(ErrorMessage.FORBIDDEN_ADMINISTRATOR);
+		}
+		return Response.success(bookService.addBook(requestDto));
 	}
 
 	@PatchMapping("/v1/books/{bookId}")
-	public ResponseEntity<Void> updateBook(
-		@PathVariable("bookId") Long bookId,
-		@RequestBody @Valid UpdateBookRequestDto requestDto
+	@ResponseStatus(HttpStatus.OK)
+	public void updateBook(
+			@AuthenticationPrincipal AuthUser user,
+			@PathVariable("bookId") Long bookId,
+			@RequestBody @Valid UpdateBookRequestDto requestDto
 	) {
+		if (!ObjectUtils.nullSafeEquals(user.getRole(), UserRole.ROLE_ADMIN)) {
+			throw new ForbiddenRequestException(ErrorMessage.FORBIDDEN_ADMINISTRATOR);
+		}
 		bookService.updateBook(bookId, requestDto);
-		return ResponseEntity.ok().build();
 	}
 
 	@PatchMapping("/v1/books/{bookId}/stock")
-	public ResponseEntity<Void> updateBookStock(
-		@PathVariable("bookId") Long bookId,
-		@RequestBody @Valid UpdateBookStockRequestDto requestDto
+	@ResponseStatus(HttpStatus.OK)
+	public void updateBookStock(
+			@AuthenticationPrincipal AuthUser user,
+			@PathVariable("bookId") Long bookId,
+			@RequestBody @Valid UpdateBookStockRequestDto requestDto
 	) {
+		if (!ObjectUtils.nullSafeEquals(user.getRole(), UserRole.ROLE_ADMIN)) {
+			throw new ForbiddenRequestException(ErrorMessage.FORBIDDEN_ADMINISTRATOR);
+		}
 		bookService.updateBookStock(bookId, requestDto);
-		return ResponseEntity.ok().build();
 	}
 
 	@DeleteMapping("/v1/books/{bookId}")
-	public ResponseEntity<Void> deleteBook(@PathVariable("bookId") Long bookId) {
+	@ResponseStatus(HttpStatus.OK)
+	public void deleteBook(
+			@AuthenticationPrincipal AuthUser user,
+			@PathVariable("bookId") Long bookId
+	) {
+		if (!ObjectUtils.nullSafeEquals(user.getRole(), UserRole.ROLE_ADMIN)) {
+			throw new ForbiddenRequestException(ErrorMessage.FORBIDDEN_ADMINISTRATOR);
+		}
 		bookService.deleteBook(bookId);
-		return ResponseEntity.ok().build();
 	}
 
 	@GetMapping("/v1/books")
-	public ResponseEntity<Page<BookResponseDto>> getAllBooks(
+	public Response<Page<BookResponseDto>> getAllBooks(
 		@RequestParam(defaultValue = "0") int page,
 		@RequestParam(defaultValue = "10") int size
 	) {
 		Pageable pageable = PageRequest.of(page, size);
-		return ResponseEntity.ok(bookService.getAllBooks(pageable));
+		return Response.success(bookService.getAllBooks(pageable));
 	}
 
 	@GetMapping("/v1/books/{bookId}")
-	public ResponseEntity<BookResponseDto> getBookById(@PathVariable("bookId") Long bookId) {
-		return ResponseEntity.ok(bookService.getBookById(bookId));
+	public Response<BookResponseDto> getBookById(@PathVariable("bookId") Long bookId) {
+		return Response.success(bookService.getBookById(bookId));
 	}
 }
